Move parsing related classes under a specific namespace
