+++ components/Components/Component.php
@@ -16,8 +16,8 @@ namespace League\Uri\Components;
 use League\Uri\Contracts\UriAccess;
 use League\Uri\Contracts\UriComponentInterface;
 use League\Uri\Contracts\UriInterface;
-use League\Uri\Encoder;
 use League\Uri\Exceptions\SyntaxError;
+use League\Uri\Parser\Encoder;
 use League\Uri\Uri;
 use Psr\Http\Message\UriInterface as Psr7UriInterface;
 use Stringable;
