+++ components/Components/UserInfo.php
@@ -17,8 +17,8 @@ use League\Uri\Contracts\AuthorityInterface;
 use League\Uri\Contracts\UriComponentInterface;
 use League\Uri\Contracts\UriInterface;
 use League\Uri\Contracts\UserInfoInterface;
-use League\Uri\Encoder;
 use League\Uri\Exceptions\SyntaxError;
+use League\Uri\Parser\Encoder;
 use Psr\Http\Message\UriInterface as Psr7UriInterface;
 use SensitiveParameter;
 use Stringable;
