+++ interfaces/QueryString.php
@@ -14,7 +14,8 @@ declare(strict_types=1);
 namespace League\Uri;
 
 use League\Uri\Exceptions\SyntaxError;
-use League\Uri\KeyValue\Converter;
+use League\Uri\Parser\Encoder;
+use League\Uri\Parser\KeyValueConverter;
 use Stringable;
 use function array_key_exists;
 use function array_keys;
@@ -55,7 +56,7 @@ final class QueryString
      */
     public static function build(iterable $pairs, string $separator = '&', int $encType = PHP_QUERY_RFC3986): ?string
     {
-        return self::buildFromPairs($pairs, Converter::fromEncodingType($encType)->withSeparator($separator));
+        return self::buildFromPairs($pairs, KeyValueConverter::fromEncodingType($encType)->withSeparator($separator));
     }
 
     /**
@@ -77,7 +78,7 @@ final class QueryString
      * @throws SyntaxError If the encoding type is invalid
      * @throws SyntaxError If a pair is invalid
      */
-    public static function buildFromPairs(iterable $pairs, Converter $converter): ?string
+    public static function buildFromPairs(iterable $pairs, KeyValueConverter $converter): ?string
     {
         $keyValuePairs = [];
         foreach ($pairs as $pair) {
@@ -107,7 +108,7 @@ final class QueryString
      */
     public static function extract(Stringable|string|bool|null $query, string $separator = '&', int $encType = PHP_QUERY_RFC3986): array
     {
-        return self::extractFromValue($query, Converter::fromEncodingType($encType)->withSeparator($separator));
+        return self::extractFromValue($query, KeyValueConverter::fromEncodingType($encType)->withSeparator($separator));
     }
 
     /**
@@ -122,7 +123,7 @@ final class QueryString
      *
      * @throws SyntaxError
      */
-    public static function extractFromValue(Stringable|string|bool|null $query, Converter $converter): array
+    public static function extractFromValue(Stringable|string|bool|null $query, KeyValueConverter $converter): array
     {
         return self::convert(self::decodePairs($converter->toPairs($query), self::PAIR_VALUE_PRESERVED));
     }
@@ -138,7 +139,7 @@ final class QueryString
      */
     public static function parse(Stringable|string|bool|null $query, string $separator = '&', int $encType = PHP_QUERY_RFC3986): array
     {
-        return self::parseFromValue($query, Converter::fromEncodingType($encType)->withSeparator($separator));
+        return self::parseFromValue($query, KeyValueConverter::fromEncodingType($encType)->withSeparator($separator));
     }
 
     /**
@@ -148,7 +149,7 @@ final class QueryString
      *
      * @return array<int, array{0:string, 1:string|null}>
      */
-    public static function parseFromValue(Stringable|string|bool|null $query, Converter $converter): array
+    public static function parseFromValue(Stringable|string|bool|null $query, KeyValueConverter $converter): array
     {
         return self::decodePairs($converter->toPairs($query), self::PAIR_VALUE_DECODED);
     }
